# admin-ui/src/gen/types/SignedDocumentUploadBodyModel.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type SignedDocumentUploadBodyModel = {
  /**
   * @description The signed PDF (multipart field `file`).
   * @type string | undefined, binary
   */
  file?: Blob;
  /**
   * @description Fallback: PDF file name (only with the base64 `file` fallback).
   * @type string | undefined
   */
  fileName?: string;
  /**
   * @description DocumentTypeDef key — MUST be an external type.
   * @type string
   */
  documentTypeKey: string;
  /**
   * @description Signature date (backdatable).
   * @type string, date-time
   */
  signedAt: string;
  /**
   * @type string | undefined
   */
  signerName?: string;
  /**
   * @type string | undefined
   */
  reference?: string;
  /**
   * @description Audience key (validated to exist when given).
   * @type string | undefined
   */
  audience?: string;
  /**
   * @type string | undefined
   */
  note?: string;
};