# openapi.admin.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/customers/{id}/signed-documents": {
      "post": {
        "operationId": "SignedDocumentsAdminController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SignedDocumentUploadBodyModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedDocumentUploadBodyModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedDocumentModel"
                }
              }
            }
          },
          "400": {
            "description": "PDF / documentTypeKey / signedAt missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "CUSTOMER_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "UNKNOWN_DOCUMENT_TYPE · DOCUMENT_TYPE_NOT_EXTERNAL · UNKNOWN_AUDIENCE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Upload an externally-signed document for a customer (multipart; base64 JSON fallback)",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "SignedDocumentsAdminController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedDocumentListResponseModel"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "List a customer’s signed documents (newest first)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/signed-documents/{id}/pdf": {
      "get": {
        "operationId": "SignedDocumentsAdminController_pdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to a fresh presigned PDF URL (time-limited)."
          },
          "401": {
            "description": "Missing/invalid admin authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "VERSION_NOT_FOUND (unknown signed document id)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "google-sso": []
          },
          {
            "admin-token": []
          }
        ],
        "summary": "Download a signed document PDF (302 → presigned URL)",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "clickwrap-server — admin API",
    "description": "Administration surface (`/admin/**`): documents & versions, publish/rollout, customers, operations (overview, history, manual acceptance, deadlines/reminders) and the dynamic entities (audiences, document types). Errors are always `{ code, message }` — the codes are documented in docs/API.md §7.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "google-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Google ID token (Google SSO; domain/allowlist checked server-side)."
      },
      "admin-token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Static admin token fallback (dev/CI; ADMIN_API_TOKEN). Optional: x-admin-user names the actor."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INVALID_STATE",
            "description": "Stable DomainErrorCode (see docs/API.md §7)."
          },
          "message": {
            "type": "string",
            "example": "A customer with externalRef \"ext-1\" and an overlapping role already exists"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "SignedDocumentUploadBodyModel": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "The signed PDF (multipart field `file`)."
          },
          "fileName": {
            "type": "string",
            "description": "Fallback: PDF file name (only with the base64 `file` fallback)."
          },
          "documentTypeKey": {
            "type": "string",
            "example": "signed-offer",
            "description": "DocumentTypeDef key — MUST be an external type."
          },
          "signedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-15T00:00:00Z",
            "description": "Signature date (backdatable)."
          },
          "signerName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "reference": {
            "type": "string",
            "example": "HubSpot deal 12345 / signed offer"
          },
          "audience": {
            "type": "string",
            "example": "customer",
            "description": "Audience key (validated to exist when given)."
          },
          "note": {
            "type": "string",
            "example": "Counter-signed on 2026-06-15."
          }
        },
        "required": [
          "documentTypeKey",
          "signedAt"
        ]
      },
      "SignedDocumentModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "sd-8a71…"
          },
          "customerId": {
            "type": "string",
            "example": "c-123"
          },
          "documentTypeKey": {
            "type": "string",
            "example": "signed-offer"
          },
          "audience": {
            "type": "string",
            "example": "customer"
          },
          "fileName": {
            "type": "string",
            "example": "signed-offer.pdf"
          },
          "contentHash": {
            "type": "string",
            "example": "sha256:…",
            "description": "SHA-256 over the PDF, computed host-side."
          },
          "fileSize": {
            "type": "number",
            "example": 20480
          },
          "signedAt": {
            "type": "string",
            "format": "date-time"
          },
          "signerName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "reference": {
            "type": "string",
            "example": "HubSpot deal 12345 / signed offer"
          },
          "note": {
            "type": "string"
          },
          "uploadedBy": {
            "type": "string",
            "example": "u-42",
            "description": "Actor who uploaded the document."
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pdfUrl": {
            "type": "string",
            "description": "Time-limited (presigned) download URL of the signed PDF."
          }
        },
        "required": [
          "id",
          "customerId",
          "documentTypeKey",
          "fileName",
          "contentHash",
          "fileSize",
          "signedAt",
          "uploadedBy",
          "uploadedAt",
          "pdfUrl"
        ]
      },
      "SignedDocumentListResponseModel": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignedDocumentModel"
            }
          }
        },
        "required": [
          "items"
        ]
      }
    }
  }
}